requirements.txt
uvicorn
psycopg2-binary
redis
temporalio
spacy
transformers
torch
pillow
requests
python-dotenv
Flask
Flask-SQLAlchemy
Flask-login
flask.ext
worldnewsapi
usermixin
authlib
flask-dance
flask-oauthlib
